requirements.txt
a2a-sdk==0.3.12
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
attrs==25.4.0
cachetools==6.2.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
decorator==5.2.1
distro==1.9.0
dnspython==2.8.0
docker==7.1.0
earthshaker==0.1.12
email-validator==2.3.0
execnet==2.1.2
executing==2.2.1
Faker==38.0.0
fastapi==0.121.1
fastapi-cli==0.0.16
fastapi-cloud-cli==0.3.1
fastapi-login==1.10.3
fastuuid==0.14.0
filelock==3.20.0
freezegun==1.5.1
frozenlist==1.8.0
fsspec==2025.10.0
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-run==0.12.0
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.1.2
idna==3.11
importlib_metadata==8.7.0
inflection==0.5.1
iniconfig==2.3.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
libcst==1.8.6
litellm==1.79.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
multidict==6.7.0
munch==4.0.0
openai==2.7.2
orjson==3.11.4
packaging==25.0
parso==0.8.5
pendulum==3.1.0
pexpect==4.9.0
pluggy==1.6.0
polyfactory==2.22.4
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic-extra-types==2.10.6
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.1
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
PyYAML-ft==8.0.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.6
rpds-py==0.28.0
rsa==4.9.1
sentry-sdk==2.44.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
SQLAlchemy-Utils==0.42.0
sqlmodel==0.0.27
sse-starlette==3.0.3
stack-data==0.6.3
starlette==0.49.3
tiktoken==0.12.0
tokenizers==0.22.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
wheel==0.45.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
# AppWorld: Commented out for Cloud Run deployment
# -e git+https://github.com/StonyBrookNLP/appworld.git@11e8d1832ffd564d3a87d30d6ad9a1e5997afda1#egg=appworld
# For Cloud Run, AppWorld data needs to be handled separately (see CLOUD_RUN_DEPLOYMENT.md)
google-cloud-storage>=2.10.0
